+++ pyblazing/pyblazing/apiv2/context.py
@@ -97,11 +97,11 @@ class blazing_allocation_mode(IntEnum):
 
 def initializeBlazing(ralId=0, networkInterface='lo', singleNode=False,
                       allocator="managed", pool=False,
-
-                      initial_pool_size=None, enable_logging=False, devices=0, config_options={}):
+                      initial_pool_size=None, enable_logging=False, 
+                      devices=0, config_options={}, logging_dir_path='blazing_log'):
 
     FORMAT='%(asctime)s|' + str(ralId) + '|%(levelname)s|||"%(message)s"||||||'
-    filename = 'pyblazing.' + str(ralId) + '.log'
+    filename = os.path.join(logging_dir_path, 'pyblazing.' + str(ralId) + '.log')
     logging.basicConfig(filename=filename,format=FORMAT, level=logging.INFO)
     workerIp = ni.ifaddresses(networkInterface)[ni.AF_INET][0]['addr']
     ralCommunicationPort = random.randint(10000, 32000) + ralId
@@ -149,9 +149,13 @@ def initializeBlazing(ralId=0, networkInterface='lo', singleNode=False,
         ralCommunicationPort,
         singleNode,
         config_options)
-    cwd = os.getcwd()
+    
+    if (os.path.isabs(logging_dir_path)):
+        log_path = logging_dir_path
+    else:
+        log_path = os.path.join(os.getcwd(), logging_dir_path)
 
-    return ralCommunicationPort, workerIp, cwd
+    return ralCommunicationPort, workerIp, log_path
 
 def getNodePartitionKeys(df, client):
     from dask import delayed
@@ -580,6 +584,37 @@ def adjust_due_to_missing_rowgroups(metadata, files):
     return metadata, new_files
 
 
+def distributed_initialize_logging_directory(client, logging_dir_path):
+
+    # lets make host_list which is a list of all the unique hosts. 
+    # This way we do the logging folder creation only once per host (server)
+    host_list = list(set([value['host'] for key,value in client.scheduler_info()["workers"].items()]))
+    initialized = {}
+    for host in host_list:
+        initialized[host]=False
+
+    dask_futures = []
+    for worker, worker_info in client.scheduler_info()["workers"].items():
+        if not initialized[worker_info['host']]:
+            dask_futures.append(
+                client.submit(
+                    initialize_logging_directory,
+                    logging_dir_path,
+                    workers=[worker]))
+            initialized[worker_info['host']] = True
+    
+    for connection in dask_futures:
+        made_dir = connection.result()
+    
+
+def initialize_logging_directory(logging_dir_path):
+    if (not os.path.exists(logging_dir_path)):
+        os.mkdir(logging_dir_path)
+        return True
+    else:
+        return False
+        
+
 class BlazingTable(object):
     def __init__(
             self,
@@ -792,7 +827,16 @@ class BlazingContext(object):
                                            default: 0.1
                                     BLAZING_DEVICE_MEM_RESOURCE_CONSUMPTION_THRESHOLD : The percent (as a decimal) of total GPU memory that the memory resource
                                             will consider to be full
+                                            NOTE: This parameter only works when used in the BlazingContext
                                             default: 0.95
+                                    BLAZING_HOST_MEM_RESOURCE_CONSUMPTION_THRESHOLD : The percent (as a decimal) of total host memory that the memory resource
+                                            will consider to be full. In the presence of several GPUs per server, this resource will be shared among all of
+                                            them in equal parts.
+                                            NOTE: This parameter only works when used in the BlazingContext
+                                            default: 0.75 
+                                    BLAZING_LOGGING_DIRECTORY : A folder path to place all logging files. The path can be relative or absolute.
+                                            NOTE: This parameter only works when used in the BlazingContext
+                                            default: 'blazing_log'                               
 
         Examples
         --------
@@ -824,13 +868,23 @@ class BlazingContext(object):
         self.finalizeCaller = ref(cio.finalizeCaller)
         self.dask_client = dask_client
         self.nodes = []
-        self.node_cwds = []
+        self.node_log_paths = []
         self.finalizeCaller = lambda: NotImplemented
         self.config_options = {}
         for option in config_options:
             self.config_options[option.encode()] = str(config_options[option]).encode() # make sure all options are encoded strings
+        
+        logging_dir_path = 'blazing_log'
+        if ('BLAZING_LOGGING_DIRECTORY' in config_options): # want to use config_options and not self.config_options since its not encoded
+            logging_dir_path = config_options['BLAZING_LOGGING_DIRECTORY']
+
+        host_memory_quota = 0.75
+        if not "BLAZING_HOST_MEM_RESOURCE_CONSUMPTION_THRESHOLD".encode() in self.config_options:
+            self.config_options["BLAZING_HOST_MEM_RESOURCE_CONSUMPTION_THRESHOLD".encode()] = str(host_memory_quota).encode()
 
         if(dask_client is not None):
+            distributed_initialize_logging_directory(self.dask_client, logging_dir_path) 
+
             if network_interface is None:
                 network_interface = 'eth0'
 
@@ -838,6 +892,13 @@ class BlazingContext(object):
             dask_futures = []
             masterIndex = 0
             i = 0
+
+            if "BLAZING_HOST_MEM_RESOURCE_CONSUMPTION_THRESHOLD" in config_options:
+                host_memory_quota = float(self.config_options["BLAZING_HOST_MEM_RESOURCE_CONSUMPTION_THRESHOLD".encode()])
+
+            #If all workers are on the same machine, the memory threshold is split between the workers, here we are assuming that there are the same number of GPUs/workers per server.
+            self.config_options["BLAZING_HOST_MEM_RESOURCE_CONSUMPTION_THRESHOLD".encode()] = str(host_memory_quota * len(set([value['host'] for key,value in self.dask_client.scheduler_info()["workers"].items()])) / len(self.dask_client.scheduler_info()["workers"])).encode()
+
             for worker in list(self.dask_client.scheduler_info()["workers"]):
                 dask_futures.append(
                     self.dask_client.submit(
@@ -850,34 +911,42 @@ class BlazingContext(object):
                         initial_pool_size=initial_pool_size,
                         enable_logging=enable_logging,
                         config_options=self.config_options,
+                        logging_dir_path=logging_dir_path,
                         workers=[worker]))
                 worker_list.append(worker)
                 i = i + 1
             i = 0
             for connection in dask_futures:
-                ralPort, ralIp, cwd = connection.result()
+                ralPort, ralIp, log_path = connection.result()
                 node = {}
                 node['worker'] = worker_list[i]
                 node['ip'] = ralIp
                 node['communication_port'] = ralPort
                 self.nodes.append(node)
-                self.node_cwds.append(cwd)
+                self.node_log_paths.append(log_path)
                 i = i + 1
 
+            
+            # need to initialize this logging independently, in case its set as a relative path 
+            # and the location from where the python script is running is different than the local dask workers
+            initialize_logging_directory(logging_dir_path) 
             # this one is for the non dask side
             FORMAT='%(asctime)s||%(levelname)s|||"%(message)s"||||||'
-            filename = 'pyblazing.log'
+            filename = os.path.join(logging_dir_path, 'pyblazing.log')
             logging.basicConfig(filename=filename,format=FORMAT, level=logging.INFO)
         else:
-            ralPort, ralIp, cwd = initializeBlazing(
+            initialize_logging_directory(logging_dir_path)
+
+            ralPort, ralIp, log_path = initializeBlazing(
                 ralId=0, networkInterface='lo', singleNode=True,
                 allocator=allocator, pool=pool, initial_pool_size=initial_pool_size,
-                enable_logging=enable_logging, config_options=self.config_options)
+                enable_logging=enable_logging, config_options=self.config_options,
+                logging_dir_path=logging_dir_path)
             node = {}
             node['ip'] = ralIp
             node['communication_port'] = ralPort
             self.nodes.append(node)
-            self.node_cwds.append(cwd)
+            self.node_log_paths.append(log_path)
 
         # NOTE ("//"+) is a neat trick to handle ip:port cases
         #internal_api.SetupOrchestratorConnection(orchestrator_host_ip, orchestrator_port)
@@ -1518,6 +1587,30 @@ class BlazingContext(object):
             else:
                 return current_table.getSlices(len(self.nodes))
 
+
+    """
+    Partition a dask_cudf DataFrame based on one or more columns.
+
+    Parameters
+    ----------
+
+    input : the dask_cudf.DataFrame you want to partition
+    by : a list of strings of the column names by which you want to partition.
+
+    Examples
+    --------
+
+    >>> bc = BlazingContext(dask_client=client)
+    >>> bc.create_table('product_reviews', "product_reviews/*.parquet")
+    >>> query_1= "SELECT pr_item_sk, pr_review_content, pr_review_sk FROM product_reviews where pr_review_content IS NOT NULL"
+    >>> product_reviews_df = bc.sql(query_1)
+    >>> product_reviews_df = bc.partition(product_reviews_df, 
+                                by=["pr_item_sk", 
+                                    "pr_review_content", 
+                                    "pr_review_sk"])
+    >>> sentences = product_reviews_df.map_partitions(create_sentences_from_reviews)
+
+    """
     def partition(self, input, by=[]):
         masterIndex = 0
         ctxToken = random.randint(0, np.iinfo(np.int32).max)
@@ -1551,7 +1644,7 @@ class BlazingContext(object):
 
 
 
-    def sql(self, query, table_list=[], algebra=None, return_futures=False, single_gpu=False, config_options={}):
+    def sql(self, query, algebra=None, return_futures=False, single_gpu=False, config_options={}):
         """
         Query a BlazingSQL table.
 
@@ -1683,8 +1776,6 @@ class BlazingContext(object):
 
         ctxToken = random.randint(0, np.iinfo(np.int32).max)
         accessToken = 0
-        if (len(table_list) > 0):
-            print("NOTE: You no longer need to send a table list to the .sql() funtion")
 
         algebra = get_plan(algebra)
 
@@ -1791,8 +1882,7 @@ class BlazingContext(object):
         """
         if not self.logs_initialized:
             self.logs_table_name = logs_table_name
-            log_files = [self.node_cwds[i] + '/RAL.' + \
-                str(i) + '.log' for i in range(0, len(self.node_cwds))]
+            log_files = [os.path.join(self.node_log_paths[i], 'RAL.' + str(i) + '.log') for i in range(0, len(self.node_log_paths))]
             dtypes = [
                 'date64',
                 'int32',
@@ -1841,9 +1931,8 @@ class BlazingContext(object):
             }
 
             for log_table_name in log_schemas:
-                log_files = [self.node_cwds[i] + '/'+log_table_name+'.' + \
-                    str(i) + '.log' for i in range(0, len(self.node_cwds))]
-
+                log_files = [os.path.join(self.node_log_paths[i], log_table_name + '.' + str(i) + '.log') for i in range(0, len(self.node_log_paths))]
+                
                 names, dtypes = log_schemas[log_table_name]
                 t = self.create_table(
                     log_table_name,